code-generator/requirements.txt
boto3>=1.35.0
python-dotenv>=1.0.0
e2b>=1.0.0
anthropic>=0.39.0
